# openapi/sdp-api.yaml
openapi: 3.0.3
info:
  title: Solana Developer Platform (SDP) API [DRAFT]
  description: |
    ⚠️ **DRAFT SPECIFICATION** — This document is a work in progress and subject to change.
    Feedback welcome. Not for production implementation yet.

    ---

    Enterprise-grade tokenization infrastructure on Solana, consolidating:
    - **Issuance** — Token-2022 creation with compliance extensions
    - **Payments** — Gasless transfers via Kora, Solana Pay integration
    - **Transactions** — Low-level signing and sending via custody providers

    ---

    ## Two Transaction Modes

    The API supports **two modes** for transaction execution:

    | Mode | Endpoint Suffix | Description |
    |------|-----------------|-------------|
    | **Prepare** | `/prepare` | Returns unsigned serialized transaction — client signs |
    | **Execute** | *(default)* | Server signs via custody provider and submits |

    ### Prepare Mode (Client-Side Signing)

    Use `/prepare` endpoints when users sign with their own wallet:

    ```
    POST /payments/transfers/prepare       → Returns unsigned tx
    POST /issuance/tokens/{id}/mint/prepare → Returns unsigned tx
    POST /transactions/prepare              → Build any unsigned tx from intent
    ```

    **Example flow:**
    ```typescript
    // 1. Get unsigned transaction from API
    const { transaction } = await fetch('/payments/transfers/prepare', {
      method: 'POST',
      body: JSON.stringify({ source, destination, amount })
...

...
    );
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key authentication
  parameters:
    environmentHeader:
      name: X-SDP-Environment
      in: header
      required: true
      schema:
        type: string
        enum: [sandbox, production]
      description: Target environment (sandbox=Devnet, production=Mainnet-Beta)
    idempotencyKey:
      name: X-SDP-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Unique key to ensure idempotent operations (required for POST/PUT/DELETE)
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of results to return
    cursorParam:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
    tokenIdPath:
      name: tokenId
      in: path
      required: true
      schema:
        type: string
        pattern: '^tok_[a-zA-Z0-9]+$'
      description: Token identifier
...
  schemas:
...
    RequestMeta:
      type: object
      properties:
        requestId:
          type: string
          example: req_abc123
      required:
        - requestId
    SuccessResponse:
      type: object
      properties:
        data:
          type: object
        meta:
          $ref: '#/components/schemas/RequestMeta'
      required:
        - data
        - meta
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - IDEMPOTENCY_CONFLICT
                - RATE_LIMIT_EXCEEDED
                - INSUFFICIENT_BALANCE
                - WALLET_NOT_ALLOWLISTED
                - WALLET_FROZEN
                - KYC_REQUIRED
                - COMPLIANCE_REJECTED
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - code
            - message
        meta:
          $ref: '#/components/schemas/RequestMeta'
      required:
        - error
        - meta
    Pubkey:
      type: string
      pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
      description: Base58-encoded Solana public key
      example: 7kbnvuGBxxj8AG9qp8Scn56muWGaRaFqxg1FsRp3PaFT
...
    Timestamp:
      type: string
      format: date-time
      example: "2026-01-23T12:00:00Z"

    # Issuance Schemas
...
    AllowlistEntry:
      type: object
      properties:
        wallet:
          $ref: '#/components/schemas/Pubkey'
        tier:
          type: string
          enum: [retail, accredited, institutional]
        jurisdiction:
          type: string
          pattern: '^[A-Z]{2}$'
          description: ISO 3166-1 alpha-2 country code
        kycReference:
          type: string
        status:
          type: string
          enum: [active, frozen, revoked]
        addedAt:
          $ref: '#/components/schemas/Timestamp'
      required:
        - wallet
        - status
    UpdateAllowlistRequest:
      type: object
      properties:
        action:
          type: string
          enum: [add, update]
        entries:
          type: array
          items:
            type: object
            properties:
              wallet:
                $ref: '#/components/schemas/Pubkey'
              tier:
                type: string
                enum: [retail, accredited, institutional]
              jurisdiction:
                type: string
              kycReference:
                type: string
            required:
              - wallet
      required:
        - action
        - entries
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_REQUEST
              message: Malformed request body
              details: {}
            meta:
              requestId: req_abc123
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or missing API key
            meta:
              requestId: req_abc123
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /issuance/tokens/{tokenId}/allowlist:
    get:
      tags: [Compliance]
      summary: Get Allowlist
      description: Retrieves the allowlist for a token
      operationId: getAllowlist
      parameters:
        - $ref: '#/components/parameters/environmentHeader'
        - $ref: '#/components/parameters/tokenIdPath'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, frozen, revoked]
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/cursorParam'
      responses:
        '200':
          description: Allowlist retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          entries:
                            type: array
                            items:
                              $ref: '#/components/schemas/AllowlistEntry'
                          nextCursor:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Compliance]
      summary: Update Allowlist
      description: Adds or updates entries in the token allowlist
      operationId: updateAllowlist
      parameters:
        - $ref: '#/components/parameters/environmentHeader'
        - $ref: '#/components/parameters/idempotencyKey'
        - $ref: '#/components/parameters/tokenIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAllowlistRequest'
      responses:
        '200':
          description: Allowlist updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          updated:
                            type: integer
                          entries:
                            type: array
                            items:
                              $ref: '#/components/schemas/AllowlistEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /issuance/tokens/{tokenId}/allowlist/{wallet}:
    delete:
      tags: [Compliance]
      summary: Remove from Allowlist
      description: Removes a wallet from the token allowlist
      operationId: removeFromAllowlist
      parameters:
        - $ref: '#/components/parameters/environmentHeader'
        - $ref: '#/components/parameters/idempotencyKey'
        - $ref: '#/components/parameters/tokenIdPath'
        - name: wallet
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Pubkey'
      responses:
        '204':
          description: Wallet removed from allowlist
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...